editoast: rename trains_schedule_base to trains_schedule
